Added a fixme for pruning.
Currently, pruning happens during the findOccurrences stage of
the algorithm. It should really happen in a separate function if
possible. Although having it here saves the overhead of querying
the tree twice, we can potentially prune more of the searchspace
by doing this somewhere else. This would offload most that work
to the SuffixTree instead of having to check for overlaps in the
MachineOutliner pass itself.