Add tail call support to the outliner.
This commit adds tail call support to the outliner. Tail calls can be outlined
with less overhead than normal calls. In x86, this can result in saving an instruction.
It also allows us to find more opportunities, because it allows us to outline returns
and other terminators.

This adds a TailCall entry to the MachineOutlinerInstrType enum in TargetInstrInfo.
If an instruction is a tail call instruction, getOutliningType returns this type.
The outliner keeps track of each return instruction's integer IDs. When a sequence
includes one as the last instruction, it considers that sequence to be a tail-callable
outlined function and includes that information when it queries the target, inserts calls,
etc.